feat: add delete confirmation message and auto delete flag (#667)

+++ cmd/minectl/delete.go
@@ -1,6 +1,11 @@
 package minectl
 
 import (
+	"bufio"
+	"fmt"
+	"os"
+	"strings"
+
 	"github.com/dirien/minectl/internal/provisioner"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
@@ -9,7 +14,8 @@ import (
 func init() {
 	deleteCmd.Flags().StringP("filename", "f", "", "Location of the manifest file")
 	deleteCmd.Flags().SetAnnotation("filename", cobra.BashCompFilenameExt, []string{"yaml"}) //nolint:errcheck
-	deleteCmd.Flags().String("id", "", "contains the server id")
+	deleteCmd.Flags().String("id", "", "Contains the server id")
+	deleteCmd.Flags().BoolP("yes", "y", false, "Automatically delete the server")
 }
 
 var deleteCmd = &cobra.Command{
@@ -46,9 +52,25 @@ func runDelete(cmd *cobra.Command, _ []string) error {
 	if err != nil {
 		return err
 	}
-	err = newProvisioner.DeleteServer()
-	if err != nil {
-		return err
+
+	yes := cmd.Flag("yes").Changed
+	if yes {
+		err = newProvisioner.DeleteServer()
+		if err != nil {
+			return err
+		}
+	} else {
+		fmt.Print("Do you want to delete the Minecraft server? [y/N]: ")
+		reader := bufio.NewReader(os.Stdin)
+		text, _ := reader.ReadString('\n')
+		text = strings.ReplaceAll(text, "\n", "")
+		if strings.Compare("y", text) == 0 {
+			err = newProvisioner.DeleteServer()
+			if err != nil {
+				return err
+			}
+		}
 	}
+
 	return nil
 }
